+++ Postal.sln
@@ -3,27 +3,10 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26730.16
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{DDA63F90-0BB7-4438-ACE7-03E71A69950B}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Build", "Build", "{F70EF96E-BECF-4500-BF12-6D982CE1C39B}"
-	ProjectSection(SolutionItems) = preProject
-		build\_ViewStart.cshtml = build\_ViewStart.cshtml
-		build\build.xml = build\build.xml
-		build\Postal.Mvc3.nuspec = build\Postal.Mvc3.nuspec
-		build\Postal.Mvc4.nuspec = build\Postal.Mvc4.nuspec
-		build\Postal.Mvc5.nuspec = build\Postal.Mvc5.nuspec
-		build\Web.config.transform = build\Web.config.transform
-	EndProjectSection
-EndProject
+
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{852A858B-75CD-423F-94EC-BA22529E7095}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Postal.Tests", "src\Postal.Tests\Postal.Tests.csproj", "{8F7B0DA6-277A-40A8-9C2A-6100112D35F6}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleSample", "src\Samples\ConsoleSample\ConsoleSample.csproj", "{2C24E942-E0E1-4116-838B-ACD25653D5B2}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ResourceSample", "src\Samples\ResourceSample\ResourceSample.csproj", "{AB26BB59-F976-4361-A041-21C7EF1F2AE5}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebSample", "src\Samples\WebSample\WebSample.csproj", "{20E5601B-354D-4393-951C-C3BCC970EAAC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Postal.AspNetCore", "src\Postal.AspNetCore\Postal.AspNetCore.csproj", "{95EE9CE4-5C7E-4D69-99C2-CB131C8CE83E}"
 EndProject
@@ -58,14 +41,6 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{DDA63F90-0BB7-4438-ACE7-03E71A69950B} = {852A858B-75CD-423F-94EC-BA22529E7095}
-		{8F7B0DA6-277A-40A8-9C2A-6100112D35F6} = {6AD69A8E-FCAB-4DF9-975A-7A9952BA8315}
-		{2C24E942-E0E1-4116-838B-ACD25653D5B2} = {DDA63F90-0BB7-4438-ACE7-03E71A69950B}
-		{AB26BB59-F976-4361-A041-21C7EF1F2AE5} = {DDA63F90-0BB7-4438-ACE7-03E71A69950B}
-		{20E5601B-354D-4393-951C-C3BCC970EAAC} = {DDA63F90-0BB7-4438-ACE7-03E71A69950B}
-		{95EE9CE4-5C7E-4D69-99C2-CB131C8CE83E} = {852A858B-75CD-423F-94EC-BA22529E7095}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {EB608C82-5210-4481-93B1-F138ECA425EC}
